Replace the RTSP output buffer with an asynchronous queue.
For RTSP buffers, instead of using an actual memory buffer (of huge
size) as output buffer, use an asynchronous queue from glib, with
GString pointers in it.

This makes not only possible to reduce the memory occupation at any
time, but also reduces the amount of work that is involved in both
appending a reply to be sent.

Additionally, asynchronous queue will work well in a multithreaded
environment to keep a thread to send data for each connection, while a
single parser prepares the replies.