Move some of the feng structure attributes in eventloop.c.
The descriptor sets and the array of listening sockets are only ever
used by the main event loop, which can be just one per process, so
there is no reason to store them in the feng structure; instead move
them in eventloop.c so that they are emitted in the .bss section.

This is useful because it improves locality of symbols, avoids
indirections to reach them, and makes it possible to ignore most user
parameters for the _foreach functions.

This required some shuffling around of functions and adding
init/cleanup functions for the eventloop, but in general reduced the
size of the code, and the heap usage.