Don't check whether g_new0 succeeded or not, for now just let glib abort.
Since there are some cases when we don't want to handle gracefully an
OOM condition, for now just ignore all the return values of g_new0 (it
will never be NULL).

Where we can expect the OOM not to act it'll be handled differently in
the future, for now simplify code.