Ensure that there is a sesion, or reply with error, and that it has the right ID.
When the session associated to the stream has an ID that differs from
the one of the session it has been paired with, reply with error to
the client.

This was already written to be done by the various methods, but is not
centralised. Note that all the methods' handling of NULL sessions was
moot because the state machine aborted first.