Replace rtsp_generate_response() with rtsp_respond().
The new function takes a RTSP_Request in input so that it can fetch
CSeq by itself, and can also check for the Timestamp header.

All responses, even negative, now get a Date and a Server header, and
Session headers are added if the request had a Session header itself.