Create a new interface to create and send RTSP responses.
Instead of always handling them as strings, handle RTSP responses as
object, similarly to what is done with requests, with an hash table
for headers.

This way it's posssible to add, remove or replace headers as needed,
and is also possible to have advanced logging capabilities when
sending the responses out.

Apply the new interface on all the methods so that it works as
intended.