Ignore Accept header during DESCRIBE.
Since we only support application/sdp as description format, there is
no reason to actually check if the client supports it. If it does not,
it'll be discard our response.

The Accept header is optional for both clients and servers (see
RFC2326, Section 12), this means that the clients might not send it
(but still support application/sdp) and servers might ignore it (and
thus the client has to check Content-Type anyway).

Once we'll support some alternative description we might parse Accept
if present, possibly in a more proper way, like described by RFC2326
Section 12.1 and by RFC2068 Section 14.1.