Change session ID handling.
Instead of keeping a 64-bit session ID saved in quadword form, make it
an arbitrary string of characters, which is made of two 32-bit random
numbers concatenated in hexadecimal form.

This way, the Session header coming from the client can be directly
compared with the saved session_id without requiring any further
processing and parsing (which is most likely more expensive in term of
computing power than a straight string comparison).

Also improve headers checking so that it can handle the situation when
a client sends an unexpected session more properly.