Define the packet only for the block where it's used.
Instead of allocating the header on the stack, and creating the packet
pointer at the outermost block, create them directly inside the if block
that is actually being used.

Also, instead of accessing the RTP_header as a local variable, access
it directly on the packet buffer, avoiding one 12 bytes memcpy later.

Generated code is the same size as before.