Clean up locking for mediathread events.
Instead of creating the list as first thing in the mediathread runner
(which might cause race conditions), add an initialisation function that
takes care of starting the mediathread runner _after_ the initialisation is
complete.

With this,t he el_mutex can also be dropped in favour of using ref/unref on
the queue itself.