Replace the schedule_list with a simpler array of session pointers.
Simplify the schedule, by removing the schedule_list structure altogether,
and replacing it with an array of pointers to RTP_session objects. This is
possible because of the old structure, the mutex is being replaced, the
valid flag is always set/clear together with the session pointer, and the
play_action function was always RTP_send_packet.

Instead of having per-entry mutexes on the array, use (for now) a single
R/W lock, which is accessed in read and write not much for the type of
access to the session but just to the actual array.

The resulting code should not only be much simpler but also much less
likely to deadlock.