Inline expand ev_io objects used for interleaved access.
Instead of using a singly-linked list (which is likely to use more memory
than just expanding the two objects), create a simple ev_io pair for the
listen instances.

Also note that the structure is already about 70KB, because of a single big
buffer, so adding a few more bytes is unlikely to change much.