Start moving handling of interleaved RTSP in its own source file.
Try to reduce the scope of all the interleaved handling code inside
rtsp_interleaved.c so that required changes can be localised.

This also allows to replace find_sctp_interleaved and find_tcp_interleaved
into a single functions since they had the same code (but were declared
static in two different source files).

With enough work it should be possible to reduce the visibility of the
structures, as well.