Move the variable down in the scope where they are used.
Instead of just using the same variables in the whole function in multiple
blocks, try to make the function more readable by creating new variable
when needed.

This is not going to change the generated code in negative way (it actually
will reduce slightly the amount of generated code since we're saving the
“->data” dereference to be executed multiple times), and makes the code
much more readable.