Refactor rtsp_read_cb().
Reduce the amount of indentation by using goto statements to jump a bit
around in the code; the result is a basically identical logic but 22 bytes
shorter in instructions.

Also rename the variables n and m to read_size and channel respectively, so
that they have some meaning when reading the code.