Merge rtsp_write_cb with rtsp_send and move it to rtsp_lowlevel.c
Instead of having a simple static callback that calls the extern rtsp_send
function, merge the two functions and move them extern to rtsp_lowlevel.c,
this way the call overhead is avoided (static or not should matter very
little given it's a callback).

This also removes the returned “written bytes” value that was being ignored
anyway.