Rewrite mediathread as two thread pools instead of an actual thread.
The result is that now instead of always having a started thread, they are
created and reused as needed for the tasks to complete.

Also rename event_buffer_low to mt_resource_read so that it actually
follows the interface.

Cleanup the mediathread.h header so that it's very quick to include, too.