Always read from the the resource instead of only reading after sending.
When we don't have data to read, and yet we don't send more, it's obvious
that we'll continue not have anything to send.

Also don't count how many frames we send. If we send more than we require
to have in cache is even a better reason for reading than having sent less
then the total cache.