Directly save the full URI for RTP sessions.
Instead of just saving the filename in the RTP sessions, and then go a long
way to find the correct value to send to the client once it's time to fill
in the RTP-Info header, just save the full URI in the structure.

Please note that the previous RTP_Session::sd_filename was only ever used
to provide the full URI. If we really need the local filename, we can
deduce it from the RTP's resource.

This fixes bug #60.