Make sure to cleanup resources properly.
Re-use the now-unused feng_free() function, move it after feng_alloc() to
have it nearer to where the stuff is allocated, make it free the whole feng
structure too and document it.

Please note that still this translates into a no-op for NDEBUG production
builds since that is only useful to make sure the memory leak tools like
valgrind don't go around bothering us about it.