Remove the feng::listeners array; instead make it static to incoming unit.
By making it static to the incoming unit it's possible to clean it up with
a simple destructor hook when building with debug enabled, and forget about
it on production builds.

This also avoids a few dereferences; even with the cleanup hook enabled,
this actually reduces the executable code generated.