Add an interface to bufferqueue to count the unseen buffers.
This actually allows to count how many buffers have yet to be seen, which
is then used by the RTP code to decide whether to read new packages or not.

With this change, the memory usage of feng does not skyrocket as soon as a
big file is played on demand, since it won't be read as quickly as the
system can, taking up gigabytes of memory.

After this change, playing the 1080p HD (MSMPEG4) version of Big Buck
Bunny (single client) tops at 10MB instead of 1GB.