Improve support for PLAY/PAUSE and seek.
Instead of just using the Range data internally, push it through the
session; use the RTSP_Range data everywhere it matters, which also means
introducing it in RTP sessions.

With this should be possible to have much finer grained timing for the RTP
packet send code, since we know _when_ we started and from which NPT.

It also supports proper PAUSE/PLAY cycle if no Range header is provided
with PLAY and puts on a foundation for the play requests edit list as
defined by RFC 2326. Right now we don't check end_time so it does not
matter to us.