Don't use the long type.
Instead of using the long type, use uint32_t. The reason for this is that
ntohl expects 32-bit parameters, and long on 64-bit machines is 64-bit.

The long type should be avoided unless needed to pass pointers, or if the
size of the type does not matter to the rest of the code.