Check for support of POSIX message queue before building the SD demuxer.
Since you could be using feng on systems where POSIX message queues are not
supported, check for its presence and make sure that it's also linked in
properly, or disable the SD demuxer altogether if mq_open is not found.

This change allows feng to link properly on FreeBSD, where otherwise it
would fail for a missing librt link. On Linux, clock_gettime is already
bringing in librt, but on FreeBSD that is not the case.