Remove entirely the *Description support.
With this change, the DESCRIBE method opens the resources repeatedly when
reading the data; this relies on the operating system's cache (which is
likely better than ours). For further optimisation we should implement a
_real_ cache of sorts, which is not what it was before.

By removing this support, we remove a huge amount of code, included the
whole MObject framework, which was just one more piece of code that could
give us headaches.