Replace the r_read function with a r_fill instead.
The r_fill function takes the consumer to fill up as an argument, and it
repeats the read until the consumer is filled up with 16 unseen
frames.

This removes the assumption 1 packet -> 1 frame, and make sures that if
both the audio and video consumers have to be filled up, it doesn't request
reading of too many packages at once.