Use g_debug for logging bufferqueue debug output.
Make the debug handling of bufferqueue debug-out (and simplify it), this
way the debug code won't be brought in the production binary (it's quite
heavy); also make the macro print the local function directly.

The tests are built with all the debugging turned on so that we have the
print out.