Fix the state machine handlers so that it actually works.
The handlers are responsible for checking how much minimum data they need
to proceed. In particular, only the basic handler and the interleaved
handler actually need an absolute amount of data (the former any amount,
the latter needs at least the four bytes to read the dynamic size
required).

Also make sure that after handling the request, the status is reset to the
start.