Add support for generic and protocol-specific response codes.
With this change we can declare some generic responses that are valid for
all the protocol supported, and others that are specific to a given
protocol.

The side-effect of this is that each protocol gets its own lookup-table,
which causes the read-only data section to get quite bigger. This will be
solved in the future, when an alternative approach, using switch/case
blocks will be optionally used.