Add support for generating ResponseCode enum from rfc822 declarations.
This replaces the RTSP_ResponseCode enumeration in netembryo, and extends
it to handle HTTP as well. Note that some response apply to all protocols,
others are protocol-specific; some use the same text as enum constant, and
others require an explicit enum to be set (for compatibility with the
previous code).