Do more sanity check when accepting a new request.
Once the request line is parsed, make sure the protocol is valid and
supported (incidentally, also make sure that the next state in the
machine is also the right one for the given protocol).

Then also make sure that the method is not invalid or unsupported, and if
it is, respond to the request directly without reading the rest of the
headers. Note: this might be debatable, because this will also make the
whole connection drop out, but the request might just be valid. Need to be
considered better!