Improve support for GIT-generation of ChangeLog.
When GIT cannot be found, don't bail out at either configure nor build
time; warn the user, and ignore the failure; if the ChangeLog is generated
empty, we can still use feng, but we cannot make a distribution.

Thus, add a distcheck-hook rule that ensures that the ChangeLog is not
empty.