Add a workaround for old, broken, locale-dependent live555.
Older versions of live555 (as used by VLC) use locale-dependent decimal
separators when preparing the NTP time for the Range header. On some
locales, this means that the separator used is comma (,) instead of dot (.)
as mandated by RFC2326 Section 3.6 (Normal Play Time).

There seem to be no incompatibility to this extension, since comma is not
used anywhere in RFC2326 as a special character, in relation to Range and
NTP, and that allows for more clients to work out of the box (namely,
Ubuntu, both 9.04 and 9.10).