Implement RTP port allocation à-la GStreamer.
Instead of trying every possible port pair, start by trying what the client
requested; if both the ports the client requested fail, then try opening
one, provided by the kernel, then use it as RTP or RTCP depending on
whether it's even or odd, and try binding the correspondent one as well.

If none can be opened this way, stop there. If only one could be open, then
get a random (kernel-assigned) port for the remaining, even if it's not
adjacent (thanks to RFC3550 this is allowed).

This brings us down to:
 - if the client provides unique ports: two calls;
 - if the client provides one good port: three calls (one may fail);
 - if the client provides no good ports: five calls (two will fail;
   one may fail).

Solves bug #20, at least partially.