Simplify even more the unicast_transport setup function.
Since the rtp_port and rtcp_port parameters are actually what the client
supplied (client_ports in the RTP-Info header), don't even consider them
(rename the parameters to make that more clear).

Reduce to the GStreamer method:

 - get a random port;
   - if even, use it for rtp, then try the one afterwards for rctp;
   - if odd, use it for rctp, then try the one before for rtp;
 - if this second bind failed, get another random port;
 - if this also failed, we can't support the transport and report so;
 - now connect it to the client's socket.

This reduce it, in the worst case, to three calls, and in the best case
two. Reducing variation in time needed is good; this gets particularly good
for localhost scenarios as the client-requested ports will obviously always
be busy.