Update to the new Ragel-based URI parser.
This removes the dependency over netembryo Url parser.

Also, a temporary URI objects is now created together with the request
unconditionally; it uses slightly more memory per request on the whole, but
we were still creating/freeing it for all requests beside OPTIONS up to
now, to verify its validity.