Update to the reduced netembyo API.
We parse the local and remote hosts (but not ports!) when creating the
client, and free them when freeing the client. This way we don't have to
store local host strings for the bound sockets.

Note: it should be possible to fetch the local_host from the bound socket
on incoming to share it between all objects, rather than having to
duplicate it per connection.