Use accept() directly rather than using neb_sock_accept().
This avoids creating one further Sock structure per client connected, and
inlines the data we need from it inside the RTSP_Client structure.

This actually brings out a need for a reduced structure for HTTP tunnelled
clients, as we now increased the size of the structure for them without
reason.

Technically, we should now consider splitting the information regarding the
socket and the session.