Don't use an illimitate number of threads for resource closing.
Instead of running an illimitate number of threads to close all the
resources, limit the number of threads in the pool to 8; when more than
eight closed connection at a time are given, they'll be queued rather than
starve up the threading system.

This avoids creating count(clients) threads if all of them disconnect at
the same time.

TODO: this should be made configurable.