Handle both the GNU-specific and the XSI-standard strerror_r function.
Thankfully, autoconf already provides a macro to test for that, we just
need to change the code to suit both options.

In the case of the GNU-specific functions, a static constant integer will
be used for the error status; this _shouldn't_ cause any overhead as the
compiler can optimise the else branch away by itself.