Bind UDP sockets to the same IP address as the client has connected to
rtsp->local_sock->local_sa is the socket address that we originally
have bound for listening, which can be 0.0.0.0. By fetching the address
that the client actually has connected to, we send packets from the
correct source address even if listening on such an address that maps
to many different addresses.

This fixes feng when running on a machine with multiple public IP addresses.