Workaround libev misdesign when hitting fileno limit.
When the limit on the number of open descriptors is hit, libev will cause
an initialisation error, that will abort feng. To work that around, perform
all the libev initialisation in-thread, and use a TLS variable to see if
it failed.

When failing the clients are rejected without further report,
unfortunately, but at least the server won't die.