Improve demuxer information again, especially extension-related.
Rather than provide four parameters to FENG_DEMUXER, provide two: shortname
and source type; the full name and the extensions are loaded from the
visible constants.

This allows to use an array for extensions, rather than a string, that
needed to be dynamically parsed. This requires us only to iterate over that
rather than call strtok (which is also non-thread-safe).