Remove support for demuxer's extension information.
All the demuxers are tested whenever we run them; since we only have three
demuxers, two for our own meta-files and one for the rest, and those two
already probe by file-extension, there is no need to run two cycles
each. Instead leave them to be probed sequentially once, and that's it.