Replace the sdp_field list with a simple append-only string.
Since the fields are only used within the context of generating the SDP
description, and that generation is the same no matter how many times you
request the same track, create the string directly on the Track structure.

This, among other, makes it faster and shorter to generate the DESCRIBE
reply, uses overall less memory, and causes less alloc/printf/free loops.