Rework a bit demuxer_sd and remove the remaining TrackInfo structure.
The mrl attribute from TrackInfo is only used by demuxer_sd to find the
actual file to load. Instead of doing that, use the private_data attribute,
and create a structure to old MRL and Queue Descriptor in a single object.

Together with this change the AVF demuxer to not use the TrackInfo
structure that is now gone for good.