Simplify handling of bound address for clients.
This fixes in a cleaner way the issue that Luca found with the source=
parameter in the SETUP setup.

Don't use the local_name as present in the Feng_Listener object, nor its
socket, to talk with the client, as that's the data for the socket that has
been bind()'ed and not the one that was accept()'ed.

Don't even save the listener address, instead add a local host string and
sockaddr object, as well as a pointer to the specific configuration. Use
the local host both to answer to SETUP as well as for SDP description.

The sockaddr structure is only needed for a bound client, and not for HTTP
tunnelled clients; the specific pointer is only used by the SCTP sockets.