Replace demuxer_sd with demuxer_sd2.
The new demuxer is almost identical in working, but the file format
changed from a completely custom language to a Key-Value Pair file as glib
allows it to be parsed.

The new format is pre-parsed, the invalid tracks are simply skipped over
and not added to the resource. This time, the order of parameters passed on
the file is not important.

The name of the track no longer depends on the name of the file, but rather
the name of the session. So for instance the following .sd file

stream
	file_name mq:///video
	encoding_name MPV
stream_end
stream
	file_name mq:///audio
	encoding_name MPA
stream_end

is equivalent to the following .sd2 file

[video]
mrl = mq:///video
encoding_name = MPV

[audio]
mrl = mq:///audio
encoding_name = MPA