Avoid keeping two (or three) copies of the socket address length per client
All the sockets for a given client (peer, local, and RTP/UDP) should use
the same addressing family, so the socket address length _have_ to be the
same for all of them. Keep thus a single copy of it.

Doublecheck the size being the same when creating the client though, with
the new assertion API.