Huge rewrite of configfile handling regarding sockets and vhosts.
There is just a single global vhost for now, but most places already look
up the data in a per-client vhost structure (which is always the global
one).

Sockets now are handled in a more generic way and there is no longer
special casing for the first socket or the others. Also it is possible to
have a number of socket options set at once:

 - IPv6 sockets actually map to both v6 proper and v4 via address mapping,
   as many other daemons do (saves a file descriptor); we already came to
   handle properly the v4/v6 decision without checking the configuration
   every time;
 - SCTP sockets are always enabled _beside_ the TCP sockets, and only after
   the equivalent TCP is open, otherwise it seems like something in Linux
   handling of SCTP causes all connections to be handled as SCTP, even if
   TCP is used by the client.