Implement a new config file parser and format.
Change the configuration syntax from a lighttpd-inspired configuration
file to a bind-inspired configuration file that seems to suit us much
better. Implement a basic options, and multiple socket and vhost sections.

To parse this file replace the old lemon-generated and hand-tweaked parser
with a flex/yacc-based reactive parser, that should be both faster and much
lighter in the amount of code produced. Since flex and yacc are repetitive
languages, use a new XML-based goldberg machine to produce the actual
source files.

[TODO: replace this goldberg machine with Perl-based generation.]

The new code parsers and makes available directly both the socket and the
vhost configuration, even though we currently make use only of the first
vhost defined (the default vhost).

Implement a --lint option that we can use to check the validity of the
configuration file without starting the server, binding ports or
anything. On the other hand, the new parser makes it difficult/impossible
to properly handle the --syslog option, so it has been dropped entirely.

Please note, currently the optional cleanup functions aren't completed, and
as such valgrind will report a number of still-reachable memory areas, all
over the parser and initialisation functions.