Use the new FFmpeg metadata API rather than the old deprecated attributes.
The new metadata API is a bit more complex to call, but covers what we need
just fine, and we don't risk to fail building when a new FFmpeg with new
API/ABI majors will be introduced.

Bump the libavformat version requirements in configure.ac to the version
that introduced the av_metadata_get() function.