sctp: use SIOCINQ if it's found in the headers.
This feature is implemented for SCTP only on Linux 2.6.37 and later. On
older kernels, the first try at using SIOCINQ ioctl() on the socket will
error out, and from there on, SIOCINQ is never used again.

TODO: ioctl() can probably error out if a client socket is dropped, it
really should just test it once per socket and not once for the whole feng
lifetime.