bufferqueue: replace bufferqueue serial with the buffer's seq_no.
With this change, the bufferqueue's serial is limited to 16-bits, which is
what is sent to the client as sequence number. For non-live streams, the
sequence number is filled like it was the old serial.

Tracks all start from the sequence number 1, rather than from a random one,
but we can also drop two attributes from the RTP_session structure.