media: implement the new demuxer-less and virtual-capable resource opening.
Change the sequence of resource opening, to check in the hash table
before doing any test, so that we can register in it virtual resources as
well. Note: this currently breaks the live resources for more than one
client at a time, as they don't register yet.

Contextually, the commit removes the Demuxer structure: the init function
is called directly, while seek, read_packet and uninit are moved on to the
Resource structure itself.