media: implement a new split handling of virtual vs stored resources.
This is a huge change that introduces the new concept of virtual resources
(such as live streaming) and moves all the stored resources support within
the libavformat "demuxer".

Parsers handling is also now folded back within that, so that there is no
reason to keep searching based on encoding names and tables.